requirements.txt
requests
python-dotenv
google-genai
google-adk
a2a-sdk